client: rename modules to match type names
